refactor: split 7 giant components 420-500 lines into focused modules (#282)
- ScriptEditor (494→190) → ScriptEditor/ + useScriptEditorController, highlight, toolbar
- ScriptReferencePanel (480→170) → ScriptReferencePanel/ + sections per category
- LabelEditDialog (474→233) → LabelEditDialog/ + fields, footer, reducer
- EditorTabBar (461→96) → EditorTabBar/ + mobile dropdown, desktop tabs, hook
- SettingsModal (476→205) → SettingsModal/ + user tab, system tab, nav
- GitLabSettingsContent (421→155) → GitLabSettingsContent/ + connected view, setup form, remove dialog, reducer
- ScriptMode page (440→152) → useScriptMode hook, useScriptModeData, ScriptModeDialogs
- VisualSystemFormContent (308→142) → 6 sibling sub-components

All barrel index.ts re-exports maintain import paths. No logic changes.
Typecheck/lint/unit tests pass. react-doctor no-giant-component: 0 warnings.
