+++ apps/frontend/src/components/ide-shared/EditorTabBar/index.ts
@@ -1,0 +1,2 @@
+export type { EditorTabBarItem } from "./EditorTabBar";
+export { EditorTabBar } from "./EditorTabBar";
