+++ apps/frontend/src/components/ide-shared/GitLabSettingsContent/index.ts
@@ -1,0 +1,1 @@
+export { GitLabSettingsContent } from "./GitLabSettingsContent";
