+++ apps/frontend/src/components/ide-shared/SettingsModal/index.ts
@@ -1,0 +1,1 @@
+export { SettingsModal } from "./SettingsModal";
