+++ apps/frontend/src/components/script-mode/ScriptEditor/index.ts
@@ -1,0 +1,2 @@
+export { ScriptEditor } from "./ScriptEditor";
+export type { ScriptEditorRef } from "./ScriptEditor";
