+++ apps/frontend/src/components/script-mode/ScriptReferencePanel/index.ts
@@ -1,0 +1,1 @@
+export { ScriptReferencePanel } from "./ScriptReferencePanel";
