+++ apps/frontend/src/components/write-mode/LabelEditDialog/index.ts
@@ -1,0 +1,2 @@
+export { LabelEditDialog } from "./LabelEditDialog";
+export type { LabelEditDialogProps } from "./LabelEditDialog";
